Hellenic Republic
Added the Hellenic Republic as a formable title for Hellenistic republics.
Also gave a young son to Christophoros Yannakakis(I mean the dude is like 55), changed that pirate's dynasty name to something more Maniot sounding(I like to thing that he chose it himself) and added a lil bit of deep lore.